+++ pyteal/ast/maybe.py
@@ -1,11 +1,15 @@
-from typing import List, Union
+from typing import Callable, List, Union, TYPE_CHECKING
 
-from pyteal.ast.multi import MultiValue
-
+from pyteal.errors import verifyProgramVersion
 from pyteal.types import TealType
 from pyteal.ir import Op
+
 from pyteal.ast.expr import Expr
 from pyteal.ast.scratch import ScratchLoad, ScratchSlot
+from pyteal.ast.multi import MultiValue
+
+if TYPE_CHECKING:
+    from pyteal.compiler import CompileOptions
 
 
 class MaybeValue(MultiValue):
@@ -17,7 +21,8 @@ class MaybeValue(MultiValue):
         type: TealType,
         *,
         immediate_args: List[Union[int, str]] = None,
-        args: List[Expr] = None
+        args: List[Expr] = None,
+        compile_check: Callable[["CompileOptions"], None] = None,
     ):
         """Create a new MaybeValue.
 
@@ -26,9 +31,28 @@ class MaybeValue(MultiValue):
             type: The type of the returned value.
             immediate_args (optional): Immediate arguments for the op. Defaults to None.
             args (optional): Stack arguments for the op. Defaults to None.
+            compile_check (optional): Callable compile check. Defaults to program version check.
+                This parameter overwrites the default program version check.
         """
+
+        # Default compile check if one is not given
+        def local_version_check(options: "CompileOptions"):
+            verifyProgramVersion(
+                minVersion=op.min_version,
+                version=options.version,
+                msg=f"{op.value} unavailable",
+            )
+
         types = [type, TealType.uint64]
-        super().__init__(op, types, immediate_args=immediate_args, args=args)
+        super().__init__(
+            op,
+            types,
+            immediate_args=immediate_args,
+            args=args,
+            compile_check=(
+                local_version_check if compile_check is None else compile_check
+            ),
+        )
 
     def hasValue(self) -> ScratchLoad:
         """Check if the value exists.
@@ -49,7 +73,7 @@ class MaybeValue(MultiValue):
     def slotOk(self) -> ScratchSlot:
         """Get the scratch slot that stores hasValue.
 
-        Note: This is mainly added for backwards compatability and normally shouldn't be used
+        Note: This is mainly added for backwards compatibility and normally shouldn't be used
         directly in pyteal code.
         """
         return self.output_slots[1]
@@ -59,7 +83,7 @@ class MaybeValue(MultiValue):
         """Get the scratch slot that stores the value or the zero value for the type if the value
         doesn't exist.
 
-        Note: This is mainly added for backwards compatability and normally shouldn't be used
+        Note: This is mainly added for backwards compatibility and normally shouldn't be used
         directly in pyteal code.
         """
         return self.output_slots[0]